pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry-core.masonry.api"

[tool.poetry]
name = "ynab-split-account"
version = "0.0.0"
authors = ["Daniel Basta <ynab@basta.info>"]
description = "Library to balance You Need A Budget (YNAB) account with counterpart in different budget"
readme = "README.md"
license = 'MIT'
packages = [{include = 'ynabsplitaccount', from = 'src'}]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = '>=3.8'
pyyaml = '>=6.0'
requests = '>=2.31'

[poetry.urls]
"Homepage" = "https://github.com/dnbasta/ynab-split-account"
"Bug Tracker" = "https://github.com/dnbasta/ynab-split-account/issues"
